Update User, Add validity check in Document

+++ ethereum/contracts/Auction.sol
@@ -1,16 +1,16 @@
 pragma solidity ^0.4.17;
 
-import './Loan.sol';
-import './Document.sol';
+import "./Loan.sol";
+import "./Document.sol";
 
 contract AuctionFactory {
     address[] deployedAuctions;
     DocumentFactory documentFactory;
-    
+
     constructor(address documentFactoryAddress) public {
         documentFactory = DocumentFactory(documentFactoryAddress);
     }
-    
+
     function createAuction(
         address newAuctionItem,
         uint256 auctionExpiryDate,
@@ -21,28 +21,30 @@ contract AuctionFactory {
             auctionExpiryDate,
             minimumBidValue
         );
-        
+
         // Lock all documents
         Loan currentLoan = Loan(newAuctionItem);
-        address[] memory documents =  currentLoan.getDocuments();
-        for (uint i = 0; i < documents.length; ++i) {
+        address[] memory documents = currentLoan.getDocuments();
+        for (uint256 i = 0; i < documents.length; ++i) {
             documentFactory.toggleDocumentLock(documents[i]);
         }
-        
+
         deployedAuctions.push(newAuction);
     }
 
     function getDeployedAuctions() public view returns (address[]) {
         return deployedAuctions;
     }
-    
-    function finalizeAuction(address currentLoanAddress, address newOwner) public {
+
+    function finalizeAuction(address currentLoanAddress, address newOwner)
+        public
+    {
         Loan loan = Loan(currentLoanAddress);
-        address[] memory documents =  loan.getDocuments();
-        
+        address[] memory documents = loan.getDocuments();
+
         // Distribute money and unlock documents
         loan.distributeAmount();
-        for (uint i = 0; i < documents.length; ++i) {
+        for (uint256 i = 0; i < documents.length; ++i) {
             documentFactory.toggleDocumentLock(documents[i]);
             documentFactory.changeDocumentOwner(documents[i], newOwner);
         }
@@ -58,7 +60,7 @@ contract Auction {
     address public currentBidder;
     address public auctionFactory;
     AuctionFactory factory;
-    
+
     constructor(
         address newAuctionItem,
         uint256 auctionExpiryDate,
@@ -111,8 +113,8 @@ contract Auction {
         loanAddress.transfer(address(this).balance);
         isActive = false;
     }
-    
-    function balance() public view returns (uint) {
+
+    function balance() public view returns (uint256) {
         return address(this).balance;
     }
 }
